Added new command functionality (setting autocomplete templates) Added simple Security to new command executing
