+++ telegram-bot-bumblebee-core/src/main/java/com/github/bumblebee/command/autocomplete/AutocompleteHandler.java
@@ -8,14 +8,11 @@ import org.springframework.stereotype.Component;
 import telegram.api.BotApi;
 import telegram.domain.Update;
 
-import java.util.Arrays;
+import java.util.HashMap;
 import java.util.Map;
-import java.util.stream.Collectors;
+import java.util.regex.Pattern;
 
 
-/**
- * Created by Misha on 14.05.2016.
- */
 @Component
 public class AutocompleteHandler extends ChainedMessageListener {
 
@@ -29,20 +26,31 @@ public class AutocompleteHandler extends ChainedMessageListener {
     public AutocompleteHandler(BotApi botApi, AutocompleteConfig config) {
         this.botApi = botApi;
         this.config = config;
-        autocompletes = config.getTemplates().stream()
-                .map(phrase -> phrase.split("/"))
-                .collect(Collectors.toMap(
-                        syllables -> syllables[0],
-                        syllables -> Arrays.copyOfRange(syllables, 1, syllables.length))
-                );
+        autocompletes = new HashMap<>();
+        for (String argument : config.getTemplates())
+            addTemplates(argument);
     }
 
-    public void addAutocompleteCommand(String argument) {
-
-        String patternKey = argument.substring(0, argument.indexOf('/'));
-        String[] patternValue = argument.replaceFirst(patternKey + "/", "").split("/");
-        autocompletes.put(patternKey, patternValue);
+    public boolean addTemplates(String argument) {
+        if (checkArgument(argument)) {
+            String patternKey = argument.substring(0, argument.indexOf('/'));
+            String[] patternValue = argument.replaceFirst(Pattern.quote(patternKey + "/"), "").split("/");
+            autocompletes.put(patternKey, patternValue);
+            return true;
+        }
+        return false;
     }
+
+    private boolean checkArgument(String argument) {
+        if (argument == null ||
+                !argument.contains("/") ||
+                argument.substring(0, argument.indexOf('/')).length() + 1 >= argument.length()) {
+
+            return false;
+        }
+        return true;
+    }
+
 
     @Override
     public boolean onMessage(Long chatId, String message, Update update) {
